Styled the nav, header and reviews section.
In the html file, the arrangement of divs in the
nav and header section was rearranged, leading to
separate containers for navbar and header; instead
of the previous method that nexted both containers
inside of one big container.

Made a change similar to the above in the reviews
section. Created a separate container to host the
image alone, then held both the image div and review
text div in another container. i.e, reviews(biggest
container) - holds Review title (section title) and
reviews. 'reviews' holds 4 'review-box'es, which
each hold a review image and review text.

The style sheet was built from the ground up to
style the nav bar as it is, style the hero section -
dividing it into two; hero header/text and subtext
on the right, and hero image on the left.
The review section was styled to hold 4 review
images/screenshots, and their quoted texts. Don't
forget the title of the section.

Alright. That was a lengthy one. I really should
commit more often. Lol. Bye for now.